feat(cli): refactor textual app structure
